Login.handlebars updated to run login and signup properly with correct class and ids, also updated login.js to run both forms correctly and redirect to dashboard page. Moved the render for dashboard from dashboardRoutes.js to homeRoutes.js
