+++ public/js/login.js
@@ -10,6 +10,28 @@ signInButton.addEventListener('click', () => {
     container.classList.remove("right-panel-active");
 });
 
+const loginFormHandler = async (event) => {
+    event.preventDefault();
+
+    // Collect login data from login form
+    const username = document.querySelector('#username-login').value.trim();
+    const password = document.querySelector('#password-login').value.trim();
+
+    if (username && password) {
+        const response = await fetch ('/api/users/login', {
+            method: 'POST',
+            body: JSON.stringify({ username, password }),
+            headers: { 'Content-Type': 'application/json' },
+        });
+
+        if (response.ok) {
+            document.location.replace('/dashboard');
+        } else {
+            alert(response.statusText);
+        }
+    }
+};
+
 const signupFormHandler = async (event) => {
     event.preventDefault();
 
@@ -33,8 +55,11 @@ const signupFormHandler = async (event) => {
             alert(response.statusText);
         }
     }
-}
+};
 
+document
+    .querySelector('.login-form')
+    .addEventListener('submit', loginFormHandler);
 
 document
     .querySelector('.signup-form')
