+++ views/homepage.handlebars
@@ -4,7 +4,7 @@
 	<div class="form-container sign-up-container">
 		<form class="signup-form">
 			<h1>Create Account</h1>
-			<input type="text" placeholder="Usermame" id="username-signup" />
+			<input type="text" placeholder="Username" id="username-signup" />
 			<input type="email" placeholder="Email" id="email-signup" />
 			<input type="password" placeholder="Password" id="password-signup" />
 			<button>Sign Up</button>
@@ -18,7 +18,6 @@
 				<a href="#" class="social"><i class="fab fa-google-plus-g"></i></a>
 				<a href="#" class="social"><i class="fab fa-linkedin-in"></i></a>
 			</div>
-			<span>or use your account</span>
 			<input type="email" placeholder="Email" />
 			<input type="password" placeholder="Password" />
 			<a href="#">Forgot your password?</a>
